Change VTCP_Check() to take a ssize_t
Since the input value is sometimes the result of a read()/write() call,
avoid truncating the ssize_t value on calling it.

+++ lib/libvarnish/vtcp.c
@@ -558,7 +558,7 @@ VTCP_check_hup(int sock)
  */
 
 int
-VTCP_Check(int a)
+VTCP_Check(ssize_t a)
 {
 	if (a == 0)
 		return (1);
